List of Modifications: - Smaller layout / removed big Images. to big 4 Mobiles - added some floatvals to avoid PHP Warnings - Add a precission to Time Calculations to meet human readable Times (amay affect caclualtion have to be tested - Debugged an issue with Sessions. Maybe related to NGINX as a Server ---> Missing Security 4 NGINX Installation. Apache recommended

+++ modules/sites_user/user01.php
@@ -35,7 +35,7 @@ if (strstr($_template->_bootstrap,'true')){
 		echo "
 		<table height='35'  border='0' cellpadding='0' cellspacing='0'>
 		<tr>";
-		$_startjahr = date("Y", intval($_user->_BeginnDerZeitrechnung));	// User - Einstellungen
+		$_startjahr = date("Y", floatval($_user->_BeginnDerZeitrechnung));	// User - Einstellungen
 		$_w_jahr	= $_time->_jahr;							// gewähltes Jahr
 		$_nextjahr	= date("Y",time());						// nächstes Jahr
 		if ($_startjahr<$_w_jahr) {
